refactor: 패키지 구조 수정 - application: 비즈니스 로직이 수행되는 클래스를 모은다. - server: 톰캣 서버측과 관련된 기능을 모은다.
