+++ src/main/java/com/woowa/hyeonsik/application/domain/User.java
@@ -1,4 +1,4 @@
-package com.woowa.hyeonsik.domain;
+package com.woowa.hyeonsik.application.domain;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
