+++ src/test/java/com/woowa/hyeonsik/application/domain/ArticleTest.java
@@ -1,5 +1,6 @@
-package com.woowa.hyeonsik.domain;
+package com.woowa.hyeonsik.application.domain;
 
+import com.woowa.hyeonsik.application.domain.Article;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
